pyproject.toml
[tool.poetry]
name = "questify-api"
version = "0.1.0"
description = "Questify App Backend"
authors = ["NikAtNight"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.25.1"
psycopg2-binary = "^2.9.9"
dj_database_url = "^2.2.0"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
django-cron = "^0.6.0"
django-cors-headers = "^4.4.0"
djangorestframework-simplejwt = "^5.3.1"
django-json-widget = "^2.0.1"
django-celery-beat = "^2.7.0"
django-import-export = "^4.1.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
